+++ packages/adapters/source/fetch.ts
@@ -6,7 +6,6 @@ import {
   trimSlashes,
 } from "dredge-common";
 import {
-  DredgeRouter,
   MiddlewareRequest,
   extractContentTypeHeader,
   getDataType,
@@ -15,6 +14,7 @@ import {
   useSuccessMiddlewares,
   useValidate,
 } from "dredge-route";
+import type { DredgeRouter } from "dredge-types";
 import { MaybePromise } from "dredge-types";
 import { ReadableStream } from "stream/web";