+++ packages/route/source/index.ts
@@ -1,5 +1,5 @@
 // export * from "./types";
-export * from "dredge-types";
+// export * from "dredge-types";
 export * from "./route";
 export { getParseFn } from "./parser";
 export * from "./route-invocation";