+++ packages/common/CHANGELOG.md
@@ -1,5 +1,7 @@
 # dredge-common
 
+## 0.5.3
+
 ## 0.5.2
 
 ### Patch Changes