+++ packages/route/CHANGELOG.md
@@ -1,5 +1,12 @@
 # dredge-route
 
+## 0.4.4
+
+### Patch Changes
+
+- Updated dependencies
+  - dredge-common@0.5.3
+
 ## 0.4.3
 
 ### Patch Changes