+++ packages/types/CHANGELOG.md
@@ -1,5 +1,7 @@
 # dredge-types
 
+## 0.4.4
+
 ## 0.4.3
 
 ### Patch Changes